Move DF templates into a template package
